% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_chdir.R
\name{chDir}
\alias{chDir}
\title{chDir}
\usage{
chDir(x, treated, control)
}
\arguments{
\item{x}{A Slinky object.}

\item{treated}{Expression data for treated samples, as
`data.frame`, `matrix`, or `SummarizedExperiment`}

\item{control}{Expression data for control samples, as
`data.frame`, `matrix`, or `SummarizedExperiment`}
}
\value{
Column matrix of characteristic direction scores for each gene.
}
\description{
Convenience wrapper to calculate Characteristic Direction Unity Vector
   based on two datasets.
   There are a few steps involved in getting the
   data formatted for the `chdirAnalysis` function. This function
   takes care of that for you.  Also, the chdirSig function is not
   exported from the package GeoDE so we copy it here to be able to
   circumvent plotting (which may not be desired for the high
   throughput applications targeted by this package).
}
