% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_gctx.R
\docType{methods}
\name{closeAll}
\alias{closeAll}
\alias{closeAll,Slinky-method}
\alias{close}
\title{closeAll}
\usage{
closeAll(x)

\S4method{closeAll}{Slinky}(x)
}
\arguments{
\item{x}{a Slinky Object}
}
\value{
None.  Called for side effect of closing
   connections.
}
\description{
Close any open HDF5 (gctx) file connections.
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
closeAll(sl)
}
