% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_clue.R
\docType{methods}
\name{clueCount}
\alias{clueCount}
\alias{clueCount,Slinky-method}
\alias{clueCount}
\title{clueCount}
\usage{
clueCount(x, endpoint = c("sigs", "cells", "genes", "perts", "plates",
  "profiles", "rep_drugs", "pcls"), where_clause = "")

\S4method{clueCount}{Slinky}(x, endpoint = c("sigs", "cells", "genes",
  "perts", "plates", "profiles", "rep_drugs", "pcls"), where_clause = "")
}
\arguments{
\item{x}{a Slinky Object}

\item{endpoint}{The endpoint to query, default is 'sigs'.}

\item{where_clause}{Optional where_clause clause.  Must be
named list (e.g. list(field='value')}
}
\value{
Count of records satisfying query
}
\description{
Wrapper for Slinky.io REST calls to retrieve record counts
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
amox_count <- clueCount(sl, where_clause = list("pert_iname" = "amoxicillin", 
                                    "cell_id" = "MCF7",
                                    "is_gold" = TRUE))
                                    
}
