% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_sumexp.R
\name{coerce}
\alias{coerce}
\alias{coerce,Slinky,SummarizedExperiment-method}
\title{as("Slinky", "SummarizedExperiment")}
\description{
Create SummarizedExperiment object from Slinky object.  Data will be 
    loaded from the GCTX file, combined with metadata from the info file, 
    and wrapped in a SummarizedExperiment object.  Note that this may take 
    a long time for the entire data set from the L1000 project.  For most 
    use cases, a subset will be desired (e.g. 
    \code{as(sl[1:987,1:50], SummarizedExperiment)}.  See the
    \code{\link{loadL1K}} method for a more flexible way to create a 
    SummarizedExperiment object based on various query parameters.
}
\examples{

# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
sumex <- as(sl[, 1:20], "SummarizedExperiment")

}
\seealso{
\code{\link{loadL1K}}
}
