% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky.R
\docType{methods}
\name{colnames}
\alias{colnames}
\alias{colnames,Slinky-method}
\alias{colnames,ANY-method}
\title{colnames}
\usage{
colnames(x, do.NULL = TRUE, prefix = "col")

\S4method{colnames}{Slinky}(x)
}
\arguments{
\item{x}{a Slinky Object}

\item{do.NULL}{Ignored (see \code{?base::colnames})}

\item{prefix}{Ignored (see \code{?base::colnames})}
}
\value{
Names of columns from gctx file
         The gctx file is an HDF5 formatted file with several
          sections (groups) containing the column and row level metadata as well
          as the expression data itself.  Note that for best performance, if a 
          subset of colnames is desired, subset the slinky object itself, not
          the colnames, to avoid loading the entire set of colnames from the the 
          gctx file.  That is, \code{names <- colnames(x[,1:50])} will be 
          considerably faster than \code{names <- colnames(x)[1:50]}.
          The \code{do.NULL} and \code{prefix} arguments from 
          \code{base::colnames} do not apply here (as the slinky object will 
          always have column names), and will be silently ignored if provided.
}
\description{
Retrieve column names from LINCS gctx datafile
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
colnames(sl[,1:5])
rownames(sl[1:5,])
}
