% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_download.R
\docType{methods}
\name{download}
\alias{download}
\alias{download,Slinky-method}
\alias{download}
\title{download}
\usage{
download(x, type = c("expression", "info", "readme"), level = 3,
  phase = 1, prompt = FALSE, verbose = FALSE)

\S4method{download}{Slinky}(x, type = c("expression", "info", "readme"),
  level = 3, phase = 1, prompt = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A Slinky object}

\item{type}{Type of file to retrieve: expression, info
(instance level), or readme file.}

\item{level}{Level of data desired (if type is expression): 2,
3 (default), 4, or 5.}

\item{phase}{What phase of data is desired, 1 of 2?  Currently
only 1 is supported. Phase 2 files may be used with this package,
but must be downloaded manually.  Phase 2 support planned for next 
update.}

\item{prompt}{Warn before downloading huge files?  Default is
FALSE.}

\item{verbose}{Do you want to know how things are going?
Default is FALSE.}
}
\value{
None
}
\description{
Convenience function to retrieve LINCS L1000 data and metadata files.
}
\note{
Most of these files are very large and may
   take many minutes to several hours to download. A fast and reliable
   connection is highly recommended.
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
sl <- new("Slinky")
\dontrun{
download(sl, type = "info")
}

}
