% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_info.R
\docType{methods}
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata,Slinky-method}
\title{get_metadata}
\usage{
get_metadata(x)

\S4method{get_metadata}{Slinky}(x)
}
\arguments{
\item{x}{a Slinky object}
}
\value{
The accessor function returns a \code{data.frame} 
containing the metadata.
}
\description{
The accessor function retrieves metadata from Slinky object.
}
\details{
As it turns out `metadata` was a poor choice for the accessor 
function because it can be masked if the user loads the 
`SummarizedExperiment` package after slinky.  So this provides 
an alternative.  Eventually `slinky::metadata` should be deprecated.
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
md <- get_metadata(sl[, 1:10])

}
