% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky.R
\docType{methods}
\name{nrow}
\alias{nrow}
\alias{nrow,Slinky-method}
\alias{nrow,ANY-method}
\alias{ncol}
\alias{ncol,Slinky-method}
\alias{ncol,ANY-method}
\title{Slinky object dimensions}
\usage{
nrow(x)

\S4method{nrow}{Slinky}(x)

ncol(x)

\S4method{ncol}{Slinky}(x)
}
\arguments{
\item{x}{an object of class Slinky}
}
\value{
number of rows or columns of current (possibly subsetted) L1000 
    data set.
}
\description{
Get the number of rows and columns in L1000 data represented by 
Slinky object.
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
ncol(sl)
nrow(sl)
}
