% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky.R
\docType{methods}
\name{rownames}
\alias{rownames}
\alias{rownames,Slinky-method}
\title{rownames}
\usage{
rownames(x, do.NULL = TRUE, prefix = "row")

\S4method{rownames}{Slinky}(x)
}
\arguments{
\item{x}{a Slinky Object}

\item{do.NULL}{Ignored (see \code{?base::rownames})}

\item{prefix}{Ignored (see \code{?base::rownames})}
}
\value{
Names of rows from gctx file
          The gctx file is an HDF5 formatted file with several
          sections (groups) containing the column and row level metadata as well
          as the expression data itself.  Note that for best performance, if a 
          subset of rownames is desired, subset the slinky object itself, not
          the rownames, to avoid loading the entire set of rownames from the the 
          gctx file.  That is, \code{names <- rownames(x[,1:50])} will be 
          faster than \code{names <- rownames(x)[1:50]}.
          The \code{do.NULL} and \code{prefix} arguments from 
          \code{base::rownames} do not apply here (as the slinky object will 
          always have row names), and will be silently ignored if provided.
}
\description{
Retrieve row names from LINCS gctx datafile
}
