% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_scoring.R
\docType{methods}
\name{rzs}
\alias{rzs}
\alias{rzs,Slinky-method}
\title{rzs}
\usage{
rzs(x, treat, control = "auto", where_clause = list(), gold = TRUE,
  inferred = TRUE, verbose = FALSE, ...)

\S4method{rzs}{Slinky}(x, treat, control = "auto", where_clause = list(),
  gold = TRUE, inferred = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{An object of class Slinky}

\item{treat}{A SummarizedExperiment containing the
treated samples,
or the pert_iname of desired perturbagen. See details.}

\item{control}{An SummarizedExperiment containing the control
samples, or the pert_iname of desired controls. Default is 'auto'.
See details.}

\item{where_clause}{If treat is a pert_iname, further query
terms may be specified here (e.g. \code{pert_type=\"trt_sh\"}).}

\item{gold}{Restrict analysis to gold instances as defined by
LINCS. Ignored if treat and control are SummarizedExperiments.}

\item{inferred}{Should the inferred (non-landmark) genes be
included in the analysis? Default is TRUE.}

\item{verbose}{Do you want to know how things are going?
Default is FALSE.}

\item{...}{Additional arguments for \code{method}.}
}
\value{
Matrix of zscore of same dimension as
    \code{treat} (or the expression matrix resulting from querying
    for \code{treat} if a pert_iname is specified).
    This function identifes same-plate controls for
    each treated sample, then converts each treated sample to robust
    z-score by subtracting the median control values and dividing by
    the (scaled) median absolute deviations.
}
\description{
Convert each sample in \code{treat} to robust zscore.
}
