% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_scoring.R
\name{.zs}
\alias{.zs}
\title{.zs}
\usage{
.zs(treat, control)
}
\arguments{
\item{treat}{a matrix of values for the treated samples}

\item{control}{a matrix of values for the control samples}
}
\value{
A vector of z-scores.
}
\description{
The .zs function provides method for calculating robust z-scores
}
