% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_package.R
\docType{package}
\name{Slinky-package}
\alias{Slinky-package}
\title{slinky: A package faciliating LINCS analysis}
\description{
The slinky package provides a wrapper to the clue.io REST API and
also utility functions for interacting with gctx (HDF5) files provided
by the LINCS program.  Future support for loading and querying data
into a MAPD instance is planned.
}
\details{
All functions are via the Slinky class which this package defines.
}
\section{Slinky functions}{

The functions ...
}

