% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky.R
\docType{methods}
\name{[,Slinky,numeric,numeric,ANY-method}
\alias{[,Slinky,numeric,numeric,ANY-method}
\alias{[,Slinky,missing,numeric,ANY-method}
\alias{[,Slinky,missing,missing,ANY-method}
\alias{[,Slinky,numeric,missing,ANY-method}
\title{Subsetting Slinky objects}
\usage{
\S4method{[}{Slinky,numeric,numeric,ANY}(x, i, j)

\S4method{[}{Slinky,missing,numeric,ANY}(x, i, j)

\S4method{[}{Slinky,missing,missing,ANY}(x, i, j)

\S4method{[}{Slinky,numeric,missing,ANY}(x, i, j)
}
\arguments{
\item{x}{A Slinky Object}

\item{i}{row index}

\item{j}{column indes
Subsets a Slinky object.  This does not touch the data on file, it simply
adjusts the index slots in the resulting Slinky object to speed up 
subsequent data operations.}
}
\value{
The subsetted Slinky object
}
\description{
Subsetting Slinky objects
}
\examples{
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
colnames(sl[,1:5])
rownames(sl[1:5,1:5])
}
