% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slinky_scoring.R
\docType{methods}
\name{rzs}
\alias{rzs}
\alias{rzs,Slinky-method}
\title{rzs}
\usage{
rzs(x, treat, control = "auto", where_clause = list(), gold = TRUE,
  inferred = TRUE, byplate = TRUE, verbose = FALSE, ...)

\S4method{rzs}{Slinky}(x, treat, control = "auto",
  where_clause = list(), gold = TRUE, inferred = TRUE,
  byplate = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{An object of class Slinky}

\item{treat}{A SummarizedExperiment containing the
treated samples,
or the pert_iname of desired perturbagen. See details.}

\item{control}{An SummarizedExperiment containing the control
samples, or the pert_iname of desired controls. Default is 'auto'.
See details.}

\item{where_clause}{If treat is a pert_iname, further query
terms may be specified here (e.g. \code{pert_type=\"trt_sh\"}).}

\item{gold}{Restrict analysis to gold instances as defined by
LINCS. Ignored if treat and control are SummarizedExperiments.}

\item{inferred}{Should the inferred (non-landmark) genes be
included in the analysis? Default is TRUE.}

\item{byplate}{Do you want to split the scores by plate? This is 
usually wise, unless you have already subsetted \code{treat} and 
\code{control} samples in such a way that plate can safely be ignored, 
or if \code{treat} and \code{control} must come from different plates 
for some reason. Default is TRUE.}

\item{verbose}{Do you want to know how things are going?
Default is FALSE.}

\item{...}{Additional arguments for \code{method}.}
}
\value{
Matrix of zscore of same dimension as
    \code{treat} (or the expression matrix resulting from querying
    for \code{treat} if a pert_iname is specified).
    This function identifes same-plate controls for
    each treated sample, then converts each treated sample to robust
    z-score by subtracting the median control values and dividing by
    the (scaled) median absolute deviations.
}
\description{
Convert each sample in \code{treat} to robust zscore.
}
\examples{
#'
# for build/demo only.  You MUST use your own key when using the slinky
# package.
user_key <- httr::content(httr::GET('https://api.clue.io/temp_api_key'),
                          as='parsed')$user_key
sl <- Slinky(user_key,
                 system.file('extdata', 'demo.gctx',
                      package='slinky'),
                 system.file('extdata', 'demo_inst_info.txt',
                     package = 'slinky'))
scores <- rzs(sl, "amoxicillin")
head(scores)
}
